refactor: create index file for mixins
