+++ packages/core/src/mixins/index.js
@@ -1,0 +1,18 @@
+import { ComponentMixin } from './component.js';
+import { PropertiesMixin } from './properties-component.js';
+import { EventsMixin } from './events-component.js';
+import { StyleMixin } from './style-component.js';
+import { TemplateMixin } from './template-component.js';
+
+/**
+ * A set of core mixins.
+ * @module MIXINS
+ * @memberof DNA
+ */
+export const MIXINS = {
+    ComponentMixin,
+    PropertiesMixin,
+    EventsMixin,
+    StyleMixin,
+    TemplateMixin,
+};
