+++ packages/custom-elements-v0/index.js
@@ -6,7 +6,8 @@
  * Evolution-based components.
  * Use with Custom Elements v0 spec.
  */
-import { mix, MIXINS } from '@dnajs/core/src/core.js';
+import { mix } from '@dnajs/core/src/core.js';
+import { MIXINS } from '@dnajs/core/src/mixins/index.js';
 import * as IDOM from '@dnajs/idom/src/lib/idom.js';
 import { registry } from '@dnajs/core/src/lib/registry.js';
 import { IDOMMixin } from '@dnajs/idom/src/mixins/idom.js';
